refactor: :recycle: change types.d to types on content types
