+++ _templates/content-type/new/package.json.t
@@ -7,12 +7,11 @@ to: packages/ripple-tide-<%= h.changeCase.paramCase(name) %>/package.json
   "description": "Ripple mappings and components for Tide <%= name %> Content type",
   "main": "dist/index.js",
   "module": "dist/index.js",
-  "types": "./types.d.ts",
   "type": "module",
   "exports": {
     ".": "./dist/index.js",
     "./component": "./src/index.vue",
-    "./types": "./types.d.ts",
+    "./types": "./src/types.ts",
     "./nuxt": "./src/nuxt.ts"
   },
   "license": "MIT",
