+++ packages/ripple-tide-event/package.json
@@ -4,12 +4,11 @@
   "description": "Ripple mappings and components for Tide Event Content type",
   "main": "dist/index.js",
   "module": "dist/index.js",
-  "types": "./types.d.ts",
   "type": "module",
   "exports": {
     ".": "./dist/index.js",
     "./component": "./src/index.vue",
-    "./types": "./types.d.ts"
+    "./types": "./src/types.ts"
   },
   "license": "MIT",
   "keywords": [],
