+++ packages/ripple-tide-grant/src/components/tide-grant-overview.vue
@@ -77,7 +77,7 @@ export default { name: 'TideGrantOverview' }
 
 <script setup lang="ts">
 import { computed } from 'vue'
-import type { TideGrantOverview } from '../../types'
+import type { TideGrantOverview } from '../types'
 import { RplButton, RplIcon, RplContent } from '@dpc-sdp/ripple-ui-core'
 
 const props =
