+++ packages/ripple-tide-landing-page/src/index.ts
@@ -1,10 +1,12 @@
 import {
   getDynamicPageComponents,
-  landingPageComponentsMapping,
-  landingPageComponentsIncludes,
   tidePageBaseMapping,
   tidePageBaseIncludes
 } from '@dpc-sdp/ripple-tide-api'
+import {
+  landingPageComponentsMapping,
+  landingPageComponentsIncludes
+} from './mapping/page-components/page-components-mapping.js'
 import type { RplTideMapping } from '@dpc-sdp/ripple-tide-api/types'
 
 const tideLandingPageModule: RplTideMapping = {
@@ -73,6 +75,7 @@ const tideLandingPageModule: RplTideMapping = {
 }
 
 export default tideLandingPageModule
+
 export {
   landingPageComponentsMapping,
   landingPageComponentsIncludes,
