+++ packages/ripple-tide-landing-page/src/index.vue
@@ -38,7 +38,7 @@
 </template>
 
 <script setup lang="ts">
-import type TideLandingPagePage from './../types'
+import type { TideLandingPagePage } from './types'
 
 defineProps<{
   page: TideLandingPagePage
