+++ packages/ripple-tide-media/src/pages/embedded-video/tide-media-embedded-video.vue
@@ -36,7 +36,7 @@ export default { name: 'TideMediaEmbeddedVideoPage' }
 </template>
 
 <script setup lang="ts">
-import type TideMediaPage from '../../../types'
+import type TideMediaPage from '../../types'
 import { RplLayout } from '@dpc-sdp/ripple-ui-core'
 import TideMediaHeader from '../../components/tide-media-header.vue'
 import TideMediaBody from '../../components/tide-media-body.vue'
