+++ tsconfig.json
@@ -23,7 +23,11 @@
     "paths": {
       "@dpc-sdp/ripple-tide-api": ["ripple-tide-api/src"],
       "@dpc-sdp/ripple-core-ui": ["ripple-core-ui/src"],
-      "@dpc-sdp/ripple-tide-event": ["ripple-tide-event/src"]
+      "@dpc-sdp/ripple-tide-event": ["ripple-tide-event/src"],
+      "@dpc-sdp/ripple-tide-grant": ["ripple-tide-grant/src"],
+      "@dpc-sdp/ripple-tide-landing-page": ["ripple-tide-landing-page/src"],
+      "@dpc-sdp/ripple-tide-media": ["ripple-tide-media/src"],
+      "@dpc-sdp/ripple-tide-publication": ["ripple-tide-publication/src"]
     }
   },
   "exclude": ["**/*.stories.ts", "**/*.stories.mdx"]
